% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrioSim_functions.R
\name{fit.risk.model.par}
\alias{fit.risk.model.par}
\title{Selecting families based on the risk model}
\usage{
fit.risk.model.par(n.ped, brks, target.snp, fam.pos, mom.tar, dad.tar, kid.tar,
  pathways, betas.e0, e.fr = NA, betas.e, pop1.frac = NA, rate.beta = NA,
  qtl = FALSE, out.put.file = NA)
}
\arguments{
\item{n.ped}{is an integer giving the number of trios to be simulated}

\item{brks}{a matrix of integers showing where the chromosomal breaks is to take place for 
each individual in the simulated trios.}

\item{target.snp}{is a vector of integers showing the row number of the target SNPs in the .bim file.}

\item{fam.pos}{is a matrix showing the chromosomal segments out of which is each target SNP 
selected for each simulated trio.}

\item{mom.tar}{is a matrix containing the target SNP genotypes of the mothers in the original data 
for simulations of a homogenous population. For simulations under population stratification it is 
a list of two matrices each containing mothers' target SNP genotypes in one of the two subpopulations.}

\item{dad.tar}{is a matrix containing the target SNP genotypes of the fathers in the original data
for simulations of a homogenous population. For simulations under population stratification it is 
a list of two matrices each containing fathers' target SNP genotypes in one of the two subpopulations.}

\item{kid.tar}{is a matrix with the target SNP genotypes of the children stacking on top of the complements in the original data
for simulations of a homogenous population. For simulations under population stratification it is 
a list of two matrices each containing children's target SNP genotypes in one of the two subpopulations.}

\item{pathways}{is a list of vectors of integers. Each vector of integers denotes the SNPs involved in a particular pathway. E.g. list(1:4,5:8)}

\item{betas.e0}{is a vector of doubles giving the beta coefficient of the logit risk model for the unexposed individuals.}

\item{e.fr}{is a double number between 0 and 1 which gives the exposure prevalence.}

\item{betas.e}{is a vector of doubles giving the beta coefficient of the logit risk model for the exposed individuals.}

\item{pop1.frac}{is a double number between 0 and 1 which gives the fraction of population 1 for a population stratification scenario.}

\item{rate.beta}{is a double number giving the log OR of disease prevalence in population 2 over that in population 1.}

\item{qtl}{is a boolean variable denoting whether a quantitative trait (qtl=TRUE) or a binary trait (qtl=FALSE) is to be simulated. For a binary trait only affected 
families will be kept. The default value is qtl=FALSE.}

\item{out.put.file}{is a character string giving the base file name for the output file. When a non-default value is given the fucntion will write the following files to 
the current working directory: 
a file with name ending with "exp.txt" containing the exposure data when exposure is involved in the risk model.
a file with name ending with "pop.txt" containing information on subpopulation membership when the simulation involves a stratified scenario.
a file with name ending with "pheno.tx" containing quantitative trait phenotype when a quantitative trait is involved.
When out.put.file is the default value NA the file names for the above three files are: exposure.txt, population.txt, phenotype.txt.}
}
\value{
The function returns a list of five elements. The first one is a matrix of integer giving the families (in terms of row number) 
selected for each chromosomal segment and each simulated trio. The second one is a matrix giving the genotypes on the 
target SNPs in the simulated trio. The third one is a vector of 0's and 1's giving the exposure status of each simulated trio when the risk model involves
exposure. The fourth element is a vector of 1's and 2's giving the subpopulation group of each simulated trio. The fifth element is a vector of doubles giving the phenotype 
values for simulations of a quantitative trait.
}
\description{
This function selects families based on the prespecified risk model. It can simulate a homogenous scenario
or a stratified scenarios with two subpopulations. For a stratified scenarion the input mom.tar, dad.tar, and
kid.tar become a list of two matrices, see Arguments for detail. When e.fr is givin rather than the default NA
the risk model can involve exposure main effects as well as gene by exposure interation. This function is parallelized
and the number of cores for parallelization is set of the total number of CPU cores minus one.
}
\examples{
tar.snp <- c(10, 203, 342, 500)
found.brks <- get.brks(N.brk=3,n.ped=1000, snp.all2, target.snp,rcmb.rate=NA)
breaks <- found.brks[[1]]
family.pos <- found.brks[[2]] 
betas <- c(-6.4, 3.2, 5.8)
pwy <- list(1:4,5:8)
fit.risk.model.par(n.ped=1000,brks=breaks,target.snp=tar.snp,fam.pos=family.position, 
mom.tar=mom.target,dad.tar=dad.target, kid.tar=kid.target, pathways=pwy, 
betas, e.fr=NA, betas,pop1.frac=0.5,rate.beta=0.3)
}

