% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrioSim_functions.R
\name{get.target.geno}
\alias{get.target.geno}
\title{Getting genotypes of the target SNPs}
\usage{
get.target.geno(input.plink.file, target.snp, snp.all2)
}
\arguments{
\item{input.plink.file}{is a vector of three character strings for the file names of the mother's
father's and child's plink base filenames. The plink files are in bed format and three files with 
extensions .bed .bim and .fam are expected for each individual's genotypes. The mothers, fathers, and 
childredn must be from the same set of trio families even though the ordering of the families can be different
for the three sets of data.}

\item{target.snp}{is a vector of integers showing the row number of the target SNPs in the .bim file}

\item{snp.all2}{is a dataframe containing list of SNPs in PLINK .bim format. Only the second column is used which contains
the rs number of the SNPs. The colname name of the second column has to be "V2".}
}
\value{
A list of three matrices is returned. The three matrices are the observed genotypes of the mothers, fathers
and children/complements at the target SNPs.
}
\description{
This function read out the genotypes of the target SNPs from the orignal data set (the data set on
which simulation is based).
}
\examples{
tar.snp <- c(10, 203, 342, 500)
assuming snps contains the data frame read from the .bim file.
get.target.geno(c('cau_mom','cau_dad','cau_kid'), tar.snp,snps)
}

