% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrioSim_functions.R
\name{glue.chr.segment.par}
\alias{glue.chr.segment.par}
\title{Splicing chromosomal segments}
\usage{
glue.chr.segment.par(input.plink.file, out.put.file, brks, sel.fam.all,
  snp.all2, pathway.all, target.snp, pop.vec = NA, no_cores = NA,
  flip = TRUE)
}
\arguments{
\item{input.plink.file}{for simulations of homogenous population, it is a vector of three character strings for the base filenames of the mother's
father's and child's plink base filenames. The plink files are in bed format and three files with 
extensions .bed .bim and .fam are expected for each individual's genotypes. The mothers, fathers, and 
childredn must be from the same set of trio families even though the ordering of the families can be different
for the three sets of data. For simulations under population stratification it is a list of two vectors. Each vector is a vector 
of three character strings for the base filenames as described above.The two vectors correspond to the two subpopulations.}

\item{out.put.file}{is a character string giving the base file name for the output file. The final file name also contains information on chromosome
number. E.g., for a base filename "trio" and for chromosome 1 the final file name is "trio1sim".}

\item{brks}{is a matrix of integers showing where the chromosomal breaks is to take place for each individual in the simulated trios.}

\item{sel.fam.all}{is a matrix of integer giving the families (in terms of row number) selected for each chromosomal segment and each simulated trio.}

\item{snp.all2}{is a dataframe containing the list of SNPs in PLINK .bim format. Two columns of the dataframe
is used: column 1 with column name "V1" containing the chromosome number and column 2 with column name "V2" containing the rs number of the SNPs.}

\item{pathway.all}{is a matrix giving the genotypes on the target SNPs in the simulated trio.}

\item{target.snp}{is a vector of integers showing the row number of the target SNPs in the .bim file.}

\item{pop.vec}{is a vector of 1's and 2's giving the subpopulation group of each simulated trio.}

\item{no_cores}{is an integer which specifies the number of CPU cores to parallelized.}
}
\value{
This function does not return values. Instead it writes PLINK files into the current directory. Each set of PLINK files contains genotype
data for one chromosome for all trios. The first third of the rows are genotypes of the mothers'. The seond third are those of the fathers' and the last third are the 
children's.
}
\description{
This function splice the chromosomal segments into "complete" trios. The spliced trio sets are written into separate plink
file chromosome by chromosome. It is parallelized and if no no_cores value is given the total number
of CPUs available will be used in the parallelization.
}
\examples{
glue.chr.segment.par(input.plink.file,out.put.file, brks,sel.fam.all,snp.all2,pathway.all,target.snp,pop.vec=NA,no_cores=NA) 
}

