% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrioSim_functions.R
\name{TrioSim}
\alias{TrioSim}
\title{TrioSim: simulate realistic trio genotypes}
\usage{
TrioSim(input.plink.file, out.put.file, fr.desire, pathways, n.ped, N.brk,
  target.snp = NA, betas.e0, betas.e1, e.fr, pop1.frac, rate.beta,
  rcmb.rate = NA, no_cores = NA, qtl = FALSE, flip = TRUE)
}
\arguments{
\item{input.plink.file}{for simulations of homogenous population, it is a vector of three character strings for the base filenames of the mother's
father's and child's plink base filenames. The plink files are in bed format and three files with 
extensions .bed .bim and .fam are expected for each individual's genotypes. The mothers, fathers, and 
childredn must be from the same set of trio families even though the ordering of the families can be different
for the three sets of data. For simulations under population stratification it is a list of two vectors. Each vector is a vector 
of three character strings for the base filenames as described above.The two vectors correspond to the two subpopulations.}

\item{out.put.file}{is a character string giving the base file name for the output file. The final file name also contains information on chromosome
number. E.g., for a base filename "trio" and for chromosome 1 the final file name is "trio1sim".}

\item{fr.desire}{is a double number giving the desired frequency of the target SNPs.}

\item{pathways}{is a list of vectors of integers. Each vector of integers denotes the SNPs involved in a particular pathway. E.g. list(1:4,5:8)}

\item{n.ped}{is an integer giving the number of trios to be simulated}

\item{N.brk}{is an integer giving the number of breaks to be picked for each chromosome.}

\item{target.snp}{is a vector of integers showing the row number of the target SNPs in the .bim file.}

\item{betas.e0}{is a vector of doubles giving the beta coefficient of the logit risk model for the unexposed individuals.}

\item{e.fr}{is a double number between 0 and 1 which gives the exposure prevalence.}

\item{pop1.frac}{is a double number between 0 and 1 which gives the fraction of population 1 for a population stratification scenario.}

\item{rate.beta}{is a double number giving the log OR of disease prevalence in population 2 over that in population 1.}

\item{rcmb.rate}{the default value is NA. rcmb.rate is a dataframe containing the recombination rates at each SNP. The ordering of the SNPs should 
be identical to that of snp.all2. It contains 4 columns with column names "V1", "V2","V3",and "V4" with the 
corresponding values for "the chromosomal number", "SNP rs number", "chromosomal position",
and "recombination rate". The recombination rate represents the maximum recombination rate in the chromosomal
region between the current SNP and the SNP above (or the first basepair of the chromosome for the first SNP on a chromosome).
When no rcmb.rate is provided the function will pick the breaking points randomly where keeping the breaking
points in between target SNPs.}

\item{no_cores}{is an integer which specifies the number of CPU cores to parallelized.}

\item{qtl}{is a boolean variable denoting whether a quantitative trait (qtl=TRUE) or a binary trait (qtl=FALSE) is to be simulated. For a binary trait only affected 
families will be kept. The default value is qtl=FALSE.}

\item{betas.e}{is a vector of doubles giving the beta coefficient of the logit risk model for the exposed individuals.}
}
\value{
this function  simulates trios and writes PLINK files into the current directory. Each set of PLINK files contains genotype
data for one chromosome for all trios. The first third of the rows are genotypes of the mothers'. The seond third are those of the fathers' and the last third are the 
children's.
}
\description{
TrioSim: simulate realistic trio genotypes
}
\examples{
baseFilename <- system.file("extdata", "pop1_4chr_dad.bed", package = "TrioSim")
baseFilename <- sub('pop1_4chr_dad.bed','',baseFilename)
input.plink.file <- c(paste(baseFilename,'pop1_4chr_mom.bed',sep=''),paste(baseFilename,'pop1_4chr_dad.bed',sep=''),
paste(baseFilename,'pop1_4chr_kid.bed',sep=''))
TrioSim(input.plink.file, 'pop1_', fr.desire=0.05,pathways=list(1:4,5:8),n.ped=1000, N.brk=3, target.snp=NA,
betas.e0=c(0,0,0), betas.e1=c(0,0,0), e.fr=NA, pop1.frac=NA, rate.beta=0,rcmb.rate, no_cores=NA)
}

